+++ eea.elasticsearch.river.rdf/.gitignore
