+++ eea.elasticsearch.river.rdf/Dockerfile
