+++ eea.elasticsearch.river.rdf/LICENSE.md
