+++ eea.elasticsearch.river.rdf/README.rst
