+++ eea.elasticsearch.river.rdf/docs/LICENSE.GPL
