+++ eea.elasticsearch.river.rdf/indexer.sh
