+++ eea.elasticsearch.river.rdf/src/logs/1/CLUSTERS-TEST1
