+++ eea.elasticsearch.river.rdf/src/logs/2/CLUSTERS-TEST
