+++ eea.elasticsearch.river.rdf/src/main/assemblies/plugin.xml
