+++ eea.elasticsearch.river.rdf/src/main/java/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: org.elasticsearch.app.Indexer
+
