+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/EEASettings.java
