+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/debug/JSONMap.java
