+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/logging/ESLogger.java
