+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/logging/Loggers.java
