+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/logging/log4j/Log4jESLogger.java
