+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/logging/slf4j/Slf4jESLogger.java
