+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/logging/support/LoggerMessageFormat.java
