+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/river/River.java
