+++ eea.elasticsearch.river.rdf/src/main/java/org/elasticsearch/app/support/ESNormalizer.java
