+++ eea.elasticsearch.river.rdf/src/main/resources/es-plugin.properties.bak
