+++ eea.elasticsearch.river.rdf/src/main/resources/log4j.properties
