+++ eea.elasticsearch.river.rdf/src/main/resources/log4j2.xml
