+++ eea.elasticsearch.river.rdf/src/test/java/org/elasticsearch/app/HarvesterTest.java
