+++ eea.elasticsearch.river.rdf/startup.sh
